Switched library loading mechanism.
The library loading mechanism has been changed to allow for multiple AppDomain's to load the library. Instead of actually loading the library, methods have been added to the Pdfium library to load and unload it using reference counting. To allow multiple AppDomain's and threads to correctly access the Pdfium library, all calls to the Pdfium library are now synchronized, because the Pdfium library is not thread safe.

Bug: #12