Expose more options for rendering PDF pages.
The Pdfium library has a number of flags that influences how a page is rendered. The ones that make sense are now exposed in the API. Also a transparency flag has been added that allows rendering PDF pages with a transparent background. This can be used when rendering to e.g. PNG files.