Take hard printer margins into account when printing.
Apparently, the Graphics object provided during printed is adjusted for the hard printer margins. This is now taken into account when printing. A setting has been added on the PdfDocument and PdfViewer that allows you to specify whether you want to account for the hard margins (default, which may loose a little bit of the page, but does preserve correct scaling) or you want to shrink the the output into the print area.

Bug: #27