Added support for rotating the PDF output.
The PdfRenderer now has a Rotation property which can be used to set the rotation. Because of limitations in either the Pdfium library or my understanding of it, rotation has been implemented through a work around (rendering to a temporary bitmap and rotating that), which will degrade performance. However, this is only done when the output is rotated.

Bug: #36