Should be able to render both annotations and form fields
Annotations flag was being used to render form fields, but not
annotations (pdf comments). Tried to avoid modifying existing method
behaviors in order to not break any existing code, but added a new
method that uses a separate bool to draw form fields and does not modify
Annotations flag so that annotations can be included in render.