Support text selection
Support selecting and copying text using the mouse cursor. PdfRenderer has a new CursorMode property to control whether to use panning or text selection.

PdfiumViewer.Demo contains a sample implementation of a right click menu. This is not implemented in PdfViewer / PdfRenderer because of localization, but it's easy to add.

PdfFile now caches the PageData for the current page so it doesn't have to reload and destroy the page data on every mouse move.